requirements.txt
pyelftools==0.32
lz4==4.4.3
